fix(dashboard): move surface helpers out of "use client" module
Every export from a `"use client"` module becomes a client reference, so
calling `parseSurfaceParam()` / `formatSurface()` from a Server Component
threw `Attempted to call ... from the server but ... is on the client`
(digest 3142508240) and rendered prod's /dashboard tree as a server
error page.

Pure helpers now live in `src/lib/surface.ts`. The client `<SurfaceFilter />`
imports them from there, and the seven dashboard server pages do the same.
