+++ src/app/dashboard/devices/page.tsx
@@ -15,7 +15,8 @@ import { deviceLabel, fmtCost, fmtNum } from "@/lib/format";
 import { PeriodSelector } from "@/components/period-selector";
 import { UnitsSelector } from "@/components/units-selector";
 import { UserFilter } from "@/components/user-filter";
-import { SurfaceFilter, parseSurfaceParam } from "@/components/surface-filter";
+import { SurfaceFilter } from "@/components/surface-filter";
+import { parseSurfaceParam } from "@/lib/surface";
 import { CostBarChart } from "@/components/charts/cost-bar-chart";
 import { DeviceCountChart } from "@/components/charts/device-count-chart";
 import { CostPerDeviceChart } from "@/components/charts/cost-per-device-chart";
