+++ src/app/dashboard/models/page.tsx
@@ -15,7 +15,8 @@ import { fmtCost, fmtNum, formatModelName, formatProvider } from "@/lib/format";
 import { PeriodSelector } from "@/components/period-selector";
 import { UnitsSelector } from "@/components/units-selector";
 import { UserFilter } from "@/components/user-filter";
-import { SurfaceFilter, parseSurfaceParam } from "@/components/surface-filter";
+import { SurfaceFilter } from "@/components/surface-filter";
+import { parseSurfaceParam } from "@/lib/surface";
 import { CostBarChart } from "@/components/charts/cost-bar-chart";
 import { ModelCountChart } from "@/components/charts/model-count-chart";
 import { CostPerModelChart } from "@/components/charts/cost-per-model-chart";
