+++ src/app/dashboard/team/page.tsx
@@ -15,7 +15,8 @@ import { parseUnit } from "@/lib/units";
 import { fmtCost, fmtNum } from "@/lib/format";
 import { PeriodSelector } from "@/components/period-selector";
 import { UnitsSelector } from "@/components/units-selector";
-import { SurfaceFilter, parseSurfaceParam } from "@/components/surface-filter";
+import { SurfaceFilter } from "@/components/surface-filter";
+import { parseSurfaceParam } from "@/lib/surface";
 import { CostBarChart } from "@/components/charts/cost-bar-chart";
 import { TeamCountChart } from "@/components/charts/team-count-chart";
 import { CostPerPersonChart } from "@/components/charts/cost-per-person-chart";
